+++ src/unpoly-migrate/classes/render_job.js
@@ -1,0 +1,6 @@
+up.migrate.renamedProperty(
+  up.RenderJob.prototype,
+  'options',
+  'renderOptions',
+  'up.RenderJob: Property { options } has been renamed to { renderOptions }'
+)