+++ src/unpoly-migrate/classes/render_result.js
@@ -1,0 +1,6 @@
+up.migrate.renamedProperty(
+  up.RenderResult.prototype,
+  'options',
+  'renderOptions',
+  'up.RenderResult: Property { options } has been renamed to { renderOptions }'
+)